stv090x, stv0910, mxl5xx: 32-bit cast for compatibility.

+++ drivers/media/dvb-frontends/stv0910.c
@@ -1488,7 +1488,7 @@ static int read_ber(struct dvb_frontend *fe, u32 *ber)
 
 	if ( p->post_bit_error.stat[0].scale == FE_SCALE_COUNTER &&
 		p->post_bit_count.stat[0].scale == FE_SCALE_COUNTER )	  
-	      *ber = p->post_bit_count.stat[0].uvalue ? p->post_bit_error.stat[0].uvalue / p->post_bit_count.stat[0].uvalue : 0;
+	      *ber = (u32)p->post_bit_count.stat[0].uvalue ? (u32)p->post_bit_error.stat[0].uvalue / (u32)p->post_bit_count.stat[0].uvalue : 0;
 
 	return 0;
 }
